Allow `MetaPromptOptimizer` to work without litellm
Extract shared LLM calling utilities (`_call_llm`, `_TokenCounter`,
`_pydantic_to_response_format`, etc.) from `mlflow/genai/discovery/utils.py`
into `mlflow/genai/utils/llm_utils.py` and use them in MetaPromptOptimizer
instead of calling `litellm.completion()` directly. This lets users run
prompt optimization with any provider supported by the gateway path
(OpenAI, Anthropic, Gemini, Mistral) without installing litellm.
